fix : Bottom Navigation && Get Group Info
