+++ app/src/main/java/org/retriever/dailypet/ui/main/AddCareFragment.kt
@@ -64,7 +64,7 @@ class AddCareFragment : BaseFragment<FragmentAddCareBinding>() {
                     is Resource.Success -> {
                         hideProgressCircular(progressCircular)
                         Toast.makeText(requireContext(), "케어가 등록되었습니다", Toast.LENGTH_SHORT).show()
-                        root.findNavController().navigate(AddCareFragmentDirections.actionAddCareFragmentToHomeFragment())
+                        root.findNavController().popBackStack()
                     }
                     is Resource.Error -> {
                         hideProgressCircular(progressCircular)
