+++ app/src/main/java/org/retriever/dailypet/ui/main/ModifyCareFragment.kt
@@ -62,7 +62,7 @@ class ModifyCareFragment : BaseFragment<FragmentModifyCareBinding>() {
                     is Resource.Success -> {
                         hideProgressCircular(progressCircular)
                         Toast.makeText(requireContext(), "케어가 수정되었습니다", Toast.LENGTH_SHORT).show()
-                        root.findNavController().navigate(ModifyCareFragmentDirections.actionModifyCareFragmentToHomeFragment())
+                        root.findNavController().popBackStack()
                     }
                     is Resource.Error -> {
                         hideProgressCircular(progressCircular)
