+++ app/src/main/res/navigation/home_navigation_graph.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/home_navigation_graph"
+    app:startDestination="@id/homeMainFragment">
+
+    <fragment
+        android:id="@+id/homeMainFragment"
+        android:name="org.retriever.dailypet.ui.main.HomeMainFragment"
+        android:label="fragment_home_main"
+        tools:layout="@layout/fragment_home_main" >
+        <action
+            android:id="@+id/action_homeMainFragment_to_statisticsFragment"
+            app:destination="@id/statisticsFragment" />
+        <action
+            android:id="@+id/action_homeMainFragment_to_modifyCareFragment"
+            app:destination="@id/modifyCareFragment" />
+        <action
+            android:id="@+id/action_homeMainFragment_to_addCareFragment"
+            app:destination="@id/addCareFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/modifyCareFragment"
+        android:name="org.retriever.dailypet.ui.main.ModifyCareFragment"
+        android:label="fragment_modify_care"
+        tools:layout="@layout/fragment_modify_care" >
+        <argument
+            android:name="petId"
+            app:argType="integer" />
+        <argument
+            android:name="careName"
+            app:argType="string" />
+        <argument
+            android:name="careId"
+            app:argType="integer" />
+    </fragment>
+    <fragment
+        android:id="@+id/addCareFragment"
+        android:name="org.retriever.dailypet.ui.main.AddCareFragment"
+        android:label="fragment_add_care"
+        tools:layout="@layout/fragment_add_care" >
+        <argument
+            android:name="petId"
+            app:argType="integer" />
+        <argument
+            android:name="petName"
+            app:argType="string" />
+    </fragment>
+    <fragment
+        android:id="@+id/statisticsFragment"
+        android:name="org.retriever.dailypet.ui.statistics.StatisticsFragment"
+        android:label="StatisticsFragment" >
+        <action
+            android:id="@+id/action_statisticsFragment_to_detailStatisticsFragment"
+            app:destination="@id/detailStatisticsFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/detailStatisticsFragment"
+        android:name="org.retriever.dailypet.ui.statistics.DetailStatisticsFragment"
+        android:label="DetailStatisticsFragment" />
+</navigation>
